Avoid piece jump when splitting.
After splitting off a piece and releasing the key you were holding,
the numbers were jumping to as if you were moving the whole piece.
It was very noticable if the computer took a bit to think. Fix by
only doing the keyUp action if you haven't dropped the piece yet.